Adding CSS flair to website.
I created a template trying to use allrecipes as an example, not
realizing that it used Bootstrap and I was going to recreate all of the
CSS of that program instead of actually understanding it. I also added
links in the form of pictures with titles for easy navigation and
formatted a title for my webpage.

[Ticket: 00001]